+++ online-store/app/Models/Article.php
@@ -10,6 +10,9 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 
+/**
+ * @property ArticleStatus $status
+ */
 class Article extends Model
 {
     use HasFactory;